adressing last PR comments
